+++ modules/typeSystem/ui/src/sap/commerce/toolset/typeSystem/folding/ItemsXmlFoldingBuilder.kt
@@ -27,11 +27,11 @@ import com.intellij.psi.util.parentOfType
 import com.intellij.psi.xml.XmlTag
 import sap.commerce.toolset.HybrisConstants
 import sap.commerce.toolset.folding.XmlFoldingBuilderEx
-import sap.commerce.toolset.settings.state.TypeSystemFoldingSettingsState
-import sap.commerce.toolset.settings.yDeveloperSettings
 import sap.commerce.toolset.typeSystem.model.*
+import sap.commerce.toolset.typeSystem.settings.TSFoldingSettings
+import sap.commerce.toolset.typeSystem.settings.state.TSFoldingSettingsState
 
-class ItemsXmlFoldingBuilder : XmlFoldingBuilderEx<TypeSystemFoldingSettingsState, Items>(Items::class.java), DumbAware {
+class ItemsXmlFoldingBuilder : XmlFoldingBuilderEx<TSFoldingSettingsState, Items>(Items::class.java), DumbAware {
 
     // it can be: EnumValue, ColumnType, but not CustomProperty
     private val valueName = "value"
@@ -64,9 +64,7 @@ class ItemsXmlFoldingBuilder : XmlFoldingBuilderEx<TypeSystemFoldingSettingsStat
         }
     }
 
-    override fun initSettings(project: Project) = project.yDeveloperSettings
-        .typeSystemSettings
-        .folding
+    override fun initSettings(project: Project) = TSFoldingSettings.getInstance().state
 
     override fun getPlaceholderText(node: ASTNode) = when (val psi = node.psi) {
         is XmlTag -> when (psi.localName) {